+++ libs/common/src/lib/interfaces/portfolio-summary.interface.ts
@@ -22,7 +22,9 @@ export interface PortfolioSummary extends PortfolioPerformance {
   grossPerformanceWithCurrencyEffect: number;
   interestInBaseCurrency: number;
   liabilitiesInBaseCurrency: number;
+  totalAssetsInBaseCurrency: number;
   totalBuy: number;
+  totalCashInBaseCurrency: number;
   totalSell: number;
   totalValueInBaseCurrency?: number;
 }